backend: modified register function

+++ backend/database/migrations/2022_12_01_165618_create_users_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->unsignedInteger('id')->unique();
+            $table->bigInteger('id')->unique();
             $table->string('name');
             $table->datetime('date_of_birth');
             $table->string('profile_picture');
